테스트 설정 변경 및 테스트 시드데이터 추가, 좋아요 관련 이벤트 및 동시성 테스트 작성 및 기능 수정
테스트 설정 변경 및 테스트 시드데이터 추가, 좋아요 관련 이벤트 및 동시성 테스트 작성

+++ src/modules/tweets/tweet.event.listener.ts
@@ -1,7 +1,7 @@
 import { Inject, Injectable, type LoggerService } from '@nestjs/common';
 import { OnEvent } from '@nestjs/event-emitter';
 import { InjectRepository } from '@nestjs/typeorm';
-import { Repository } from 'typeorm';
+import { EntityManager, Repository } from 'typeorm';
 import { TweetLike } from './entities/tweet-like.entity';
 import { TweetLikeEvent } from './events/tweet-like.event';
 import { WINSTON_MODULE_PROVIDER } from 'nest-winston';
@@ -11,22 +11,63 @@ export class TweetEventListener {
   constructor(
     @Inject(WINSTON_MODULE_PROVIDER) private readonly logger: LoggerService,
     @InjectRepository(TweetLike) private readonly repo: Repository<TweetLike>,
+    private readonly entityManager: EntityManager,
   ) {}
+
+  private jobMap = new Map<string, (() => Promise<void>)[]>();
+  private processMap = new Map<string, boolean>();
+
+  private async processEvent(name: string) {
+    const jobs = this.jobMap.get(name);
+    if (this.processMap.get(name) || !jobs || jobs.length === 0) return;
+
+    this.processMap.set(name, true);
+    const job = jobs.shift();
+
+    if (!job) {
+      this.processMap.set(name, false);
+      return;
+    }
+
+    try {
+      await job();
+    } finally {
+      this.processMap.set(name, false);
+      setImmediate(() => this.processEvent(name));
+    }
+  }
 
   @OnEvent('tweet.liked')
   async handleTweetLikedEvent(event: TweetLikeEvent) {
-    try {
-      await this.repo.save({ tweet: { id: event.tweetId }, user: { id: event.userId } });
-    } catch (error) {
-      this.logger.warn('Event Fail (tweet.liked)', error);
-    }
+    const name = 'tweet.liked';
+
+    if (!this.jobMap.has(name)) this.jobMap.set(name, []);
+
+    this.jobMap.get(name)!.push(async () => {
+      try {
+        await this.repo.insert({ tweetId: event.tweetId, userId: event.userId });
+      } catch (error) {
+        this.logger.warn(`Event Fail (${name})`, error);
+      }
+    });
+
+    await this.processEvent(name);
   }
 
   @OnEvent('tweet.unliked')
   async handleTweetUnlikedEvent(event: TweetLikeEvent) {
-    await this.repo.delete({
-      tweet: { id: event.tweetId },
-      user: { id: event.userId },
+    const name = 'tweet.unliked';
+
+    if (!this.jobMap.has(name)) this.jobMap.set(name, []);
+
+    this.jobMap.get(name)!.push(async () => {
+      try {
+        await this.repo.delete({ tweetId: event.tweetId, userId: event.userId });
+      } catch (error) {
+        this.logger.warn(`Event Fail (${name})`, error);
+      }
     });
+
+    await this.processEvent(name);
   }
 }
